pyproject.toml
[project]
name = "ozon-bi"
version = "0.1.0"
description = "OZON Business Intelligence and Automated Pricing System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=20.7",
    "httpx>=0.26.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "apscheduler>=3.10.4",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.3",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
